opentelemetry/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging